backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-docx>=1.1.0
pydantic>=2.5.0
# 注意：pydantic需要预编译的wheel包，使用 --only-binary :all: pydantic 安装

